pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-manager"
version = "0.1.0"
description = "TUI Manager for Claude Code Agents and Skills"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Malcolm" }
]
keywords = ["claude", "agents", "tui", "terminal", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: User Interfaces",
]

dependencies = [
    "textual>=0.75.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "jsonschema>=4.0",
    "typer>=0.9.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "textual-dev>=1.0",
]

[project.scripts]
agent-manager = "agent_manager.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_manager"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
